scrubbed some remaining bugs
So I just finished scrubbing out some annoying bugs that removed all null pointer exceptions. Kotlin doesn't do null pointer exceptions, its more dignified than that! Its perfect. It's official though, the entire project has been converted to kotlin. This is no longer just a test branch, this is Jolt-Sphere, I think I'm going to merge this into the master branch now, but I'm also going to branch off of Matthew's branch for the java based HTML version, it's going to be beautiful, thank you for your time and thank you jetbrains for this gift that is kotlin, I'll never be able to happily work with java ever again. Good night world!!!

+++ core/src/org/joltsphere/scenes/Scene5.kt
@@ -62,7 +62,7 @@ class Scene5(internal val game: JoltSphereMain) : Screen {
         mapHeight = mapProp.get("height", Int::class.java) as Int * 320
         MapBodyBuilder.buildShapes(map, ppm, world, "terrain")
 
-        streamBeam = StreamBeamTowedPlayer(world, 450, (mapHeight / 1.8f).toInt(), Color.RED)
+        streamBeam = StreamBeamTowedPlayer(world, 450f, (mapHeight / 1.8f).toFloat(), Color.RED)
         otherPlayer = MountainClimbingPlayer(world, 500f, (mapHeight / 1.8f).toInt().toFloat(), Color.BLUE)
 
         val dDef = DistanceJointDef()
